2.3.2
git-svn-id: https://plugins.svn.wordpress.org/zingiri-web-shop/trunk@469013 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ fwkfor/classes/sub.file_multiple.class.php
@@ -4,12 +4,14 @@ class file_multipleZfSubElement extends zfSubElement {
 	function postSave($id=0)
 	{
 		$prefix=$this->element->input['upload_file_key'];
-		$product_dir=@constant($this->element->populated_value['element_'.$this->elementid.'_'.($this->subid+1)]);
-		$totalFiles=count($this->element->input['loaded_files'])+count($this->element->input['new_images'])-count($this->element->input['delimage']);
+		$product_dir=defined('APHPS_DATA_DIR') ? APHPS_DATA_DIR : constant($this->element->populated_value['element_'.$this->elementid.'_'.($this->subid+1)]);
+		$totalFiles=isset($this->element->input['loaded_files']) ? count($this->element->input['loaded_files']) : 0;
+		$totalFiles=$totalFiles+(isset($this->element->input['new_images']) ? count($this->element->input['new_images']) : 0);
+		$totalFiles=$totalFiles-(isset($this->element->input['delimage']) ? count($this->element->input['delimage']) : 0);
 		$picid=$id;
 
 		//delete images if required
-		if (count($this->element->input['delimage'])>0) {
+		if (isset($this->element->input['delimage']) && count($this->element->input['delimage'])>0) {
 			foreach ($this->element->input['delimage'] as $imageid) {
 				unlink($product_dir.'/'.$prefix.'__'.$imageid);
 			}
@@ -22,7 +24,7 @@ class file_multipleZfSubElement extends zfSubElement {
 		$db=new aphpsDb();
 		$db->updateRecord($this->element->entity,array('ID' => $id),array($column => $link));
 		$this->ext=$this->int=$link;
-		
+
 		return true;
 	}
 
@@ -31,21 +33,25 @@ class file_multipleZfSubElement extends zfSubElement {
 		$i=$this->subid;
 		$xmlf=$this->xmlf;
 
-		$constant_dir=$this->element->populated_value['element_'.$this->elementid.'_'.($this->subid+1)];
-		$product_dir=@constant($this->element->populated_value['element_'.$this->elementid.'_'.($this->subid+1)]);
+		$constant_dir=isset($this->element->populated_value['element_'.$this->elementid.'_'.($this->subid+1)]) ? $this->element->populated_value['element_'.$this->elementid.'_'.($this->subid+1)] : '';
+		$product_dir=defined('APHPS_DATA_DIR') ? APHPS_DATA_DIR : constant($constant_dir);
 		if (strstr($this->element->populated_value['element_'.$this->elementid.'_'.$this->subid],'__')) list($filePrefix,$fileName)=explode('__',$this->element->populated_value['element_'.$this->elementid.'_'.$this->subid]);
-		else $filePrefix=$fileName='';
-		if (empty($filePrefix)) $filePrefix=create_sessionid(16,1,36);
+		elseif ($this->element->populated_value['element_'.$this->elementid.'_'.$this->subid]) {
+			$filePrefix=$this->element->populated_value['element_'.$this->elementid.'_'.$this->subid];
+			$fileName='';
+		} else $filePrefix=$fileName='';
+		if (empty($filePrefix)) $filePrefix=$this->createRandomCode(16);
 
-		$picid=is_numeric($_GET['id']) ? $_GET['id'] : 0;
+		$picid=(isset($_GET['id']) && is_numeric($_GET['id'])) ? $_GET['id'] : 0;
 
-		$field_markup.=loadJavascript(ZING_APPS_PLAYER_URL . 'js/' . APHPS_JSDIR . '/ajaxupload.js');
-		$field_markup.=loadJavascript(ZING_APPS_PLAYER_URL . 'js/' . APHPS_JSDIR . '/fileupload.jquery.js');
-
-		$field_markup.='<input type="button" id="upload_file_button" value="'.zurl('Upload').'" />';
-		$field_markup.='<input type="hidden" id="upload_file_dir" value="'.$constant_dir.'" />';
-		$field_markup.='<input type="hidden" name="upload_file_key" id="upload_file_key" value="'.$filePrefix.'">';
-
+		if ($this->mode != 'view') {
+			$field_markup.=loadJavascript(ZING_APPS_PLAYER_URL . 'js/' . APHPS_JSDIR . '/ajaxupload.js');
+			$field_markup.=loadJavascript(ZING_APPS_PLAYER_URL . 'js/' . APHPS_JSDIR . '/fileupload.jquery.js');
+			$field_markup.='<input type="button" id="upload_file_button" value="'.zurl('Upload').'" />';
+			$field_markup.='<input type="hidden" id="upload_file_dir" value="'.$constant_dir.'" />';
+			$field_markup.='<input type="hidden" name="upload_file_key" id="upload_file_key" value="'.$filePrefix.'">';
+		}
+		
 		$imgs=array();
 		$field_markup.='<ul id="uploaded_files">';
 		if ($handle=opendir($product_dir)) {
@@ -61,12 +67,16 @@ class file_multipleZfSubElement extends zfSubElement {
 			asort($imgs);
 			foreach ($imgs as $img) {
 				$field_markup.='<li id="'.$img.'" style="position:relative;clear:both">';
-				$field_markup.='<p>'.$img.'</p>';
-				$field_markup.='<a onclick="wsDeleteFile(\''.$img.'\');">';
-				$field_markup.='<img style="position:absolute;right:-16px;top:0px;" src="'.ZING_APPS_PLAYER_URL.'images/delete.png" height="16px" width="16px" />';
-				$field_markup.="</a>";
-				$field_markup.='<input name="loaded_files[]" type="hidden" value="'.$img.'" />';
-				$field_markup.='</li>';
+				if ($this->mode=='view') {
+					$field_markup.='<a href="'.$img.'">'.$img.'</a>';	
+				} else {
+					$field_markup.='<p>'.$img.'</p>';
+					$field_markup.='<a onclick="wsDeleteFile(\''.$img.'\');">';
+					$field_markup.='<img style="position:absolute;right:-16px;top:0px;" src="'.ZING_APPS_PLAYER_URL.'images/delete.png" height="16px" width="16px" />';
+					$field_markup.="</a>";
+					$field_markup.='<input name="loaded_files[]" type="hidden" value="'.$img.'" />';
+					$field_markup.='</li>';
+				}
 			}
 		}
 		$field_markup.='</ul><div style="clear:both"></div>';
